requirements.txt
celery
kombu[sqlalchemy]
fastapi
uvicorn
openai
faiss-cpu
pypdf
python-multipart
requests
jinja2
PyPDF2
uvicorn[standard]
websockets
wsproto
fuzzywuzzy
aiohttp
aiosqlite
redis 
loguru 
slowapi 
aioredis 
tenacity 
sqlalchemy[asyncio]>=2.0.0
watchdog
portalocker
pypdf
psutil
scipy
lightrag-hku
docx
greenlet
aiosmtplib
honcho==1.1.0
cryptography
yt_dlp>=2025.12.8
google-generativeai
google-genai
instaloader
facebook-scraper
twscrape
beautifulsoup4
pillow
twitter-scraper
werkzeug
oauth2
tweepy
gallery-dl
eventlet
tiktoken
bcrypt
ipinfo
moviepy
pytubefix
pytube
moviepy
numpy>=2.1.0
passlib[argon2]
apify-client